+++ README.md
@@ -1,11 +1,11 @@
 # Booklist-App
 
->  Booklist-App Project For Microverse By Bwalya Nonde, Patrick Ndimukulaga and Sajjad Rajaee
+>  Booklist-App Project For Microverse By Bwalya Nonde.
 
 
 > This is a simple web develoment project for microverse created using simple HTML(Hyper text mark-up   Language.) and CSS(Cascading Style Sheet) which is used for styling or decorating the webpage. And JavaScript which is used for making the webpage responsive.
 
-![screenshot](awe-some.PNG)
+![screenshot](modules.png)
 
 
 ## Built With
