+++ modules/time.js
@@ -1,0 +1,5 @@
+import { DateTime } from '../node_modules/luxon/build/es6/luxon.js';
+
+const Date = DateTime.now().toHTTP();
+
+export default Date;
